% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsurvROC.R
\name{lsurvROC}
\alias{lsurvROC}
\title{lsurvROC}
\usage{
lsurvROC(
  dat.long,
  dat.short,
  cutoff.type.basis,
  sens.type.basis,
  covariate1 = 1,
  covariate2 = 1,
  tau,
  time.window,
  nResap = 200,
  newdata = NULL,
  nknot = NULL,
  monotone.method = "both",
  tol = 1000
)
}
\arguments{
\item{dat.long}{a data frame with long data format. It should contain subject id (id), longitudinal marker (Xt), measurement time (vtime)}

\item{dat.short}{a data frame. It should contain unique subject id (id), observed event time (Y), event indicator (delta, 1 = observed event, 0 = censored), covariates}

\item{cutoff.type.basis}{basis function to estimate marker threshold ("FP" = fractional polynomial, "bs" = "B splines", "linear" = linear basis)}

\item{sens.type.basis}{basis function to estimate sensitivity levels ("FP" = fractional polynomial, "bs" = "B splines", "linear" = linear basis)}

\item{covariate1}{a vector of name of covariates to estimate marker thresholds at level tau(s)}

\item{covariate2}{a vector of name of covariates to estimate sensitivity levels}

\item{tau}{the quantile(s) of marker threshold to be estimated}

\item{time.window}{a numeric value. This value will be used to specify the future time point to perform the evaluation, meaning vtime + time.window}

\item{nResap}{a numeric value to perform perturbation resampling. The default value = 200. Larger value can generate more accurate ASE value}

\item{nknot}{if basis function is B splines, need to specify the number of knot}

\item{monotone.method}{user specified method for monotonization.}

\item{tol}{a constant used for removing unconverged results}

\item{show_plots}{if TRUE then the time-varying coefficients plot will be displayed}
}
\description{
It runs flexible regression models to estimate biomarker thresholds at selected quantile(s) and the corresponding sensitivities.
It returns model results
and plots the time-varying coefficients with 95\% confidence interval.
}
\examples{
  data(example_data)
  head(example_data$data.long)
  head(example_data$data.short)
  
  res <- lsurvROC(dat.long = example_data$data.long, 
    dat.short = example_data$data.short,
    cutoff.type.basis = "FP",
    sens.type.basis = "FP", 
    covariate1 = c("Z", "Zcont"), 
    covariate2 = c("Z", "Zcont"), 
    tau = c(0.7, 0.8, 0.9), 
    time.window = 1, 
    nResap = 50, 
    show_plots = TRUE
    )
}
